Fixes #7543: consolidate test builders
