+++ python/subprocess-via-runner-exemptions.json
@@ -14,5 +14,9 @@
   {
     "file": "tests/skills/_structure_review_specialized.py",
     "reason": "Skill-structure specialized port for make test-review-structure; intentionally spawns python3 cli.py subprocesses to exercise real CLI behavior under pytest."
+  },
+  {
+    "file": "tests/support/foundation.py",
+    "reason": "Shared pytest fixture module intentionally creates disposable git repositories and invokes cli.py subprocesses to exercise integration behavior."
   }
 ]
