+++ python/tests/lint/__init__.py
@@ -1,0 +1,1 @@
+"""Lint-rule test package."""
