+++ python/tests/lint/test_lint_flat_tests.py
@@ -7,6 +7,7 @@ from pathlib import Path
 import pytest
 
 from larch.lint.lint_flat_tests import main
+from tests.lint.conftest import lint_runner
 
 
 def write(root: Path, rel: str, text: str = "x\n") -> None:
@@ -15,9 +16,7 @@ def write(root: Path, rel: str, text: str = "x\n") -> None:
     _ = path.write_text(text, encoding="utf-8")
 
 
-def run(root: Path, capsys: pytest.CaptureFixture[str]) -> tuple[int, str]:
-    rc = main(["--root", str(root)])
-    return rc, capsys.readouterr().err
+run = lint_runner(main)
 
 
 def test_clean_tree_without_root_tests_passes(tmp_path: Path, capsys: pytest.CaptureFixture[str]) -> None:
