+++ python/tests/lint/test_lint_shared_convention_regex.py
@@ -4,6 +4,7 @@ from pathlib import Path
 from typing import Protocol
 
 from larch.lint import lint_shared_convention_regex as lscr
+from tests.lint.conftest import write_project as _write_project
 
 
 class CaptureResult(Protocol):
@@ -13,14 +14,6 @@ class CaptureResult(Protocol):
 
 class CaptureFixture(Protocol):
     def readouterr(self) -> CaptureResult: ...
-
-
-def _write_project(root: Path, *, files: dict[str, str]) -> None:
-    python_dir = root / "python"
-    for relpath, source in files.items():
-        path = python_dir / relpath
-        path.parent.mkdir(parents=True, exist_ok=True)
-        _ = path.write_text(source, encoding="utf-8")
 
 
 def _module(body: str) -> str:
