+++ python/tests/lint/test_lint_skill_description_length.py
@@ -4,20 +4,13 @@ from pathlib import Path
 from typing import TYPE_CHECKING
 
 from larch.lint.lint_skill_description_length import main
+from tests.lint.conftest import lint_runner, write_skill
 
 if TYPE_CHECKING:
     import pytest
 
 
-def write_skill(root: Path, rel: str, body: str) -> None:
-    path = root / rel
-    path.parent.mkdir(parents=True, exist_ok=True)
-    _ = path.write_text(body, encoding="utf-8")
-
-
-def run(root: Path, capsys: pytest.CaptureFixture[str]) -> tuple[int, str]:
-    rc = main(["--root", str(root)])
-    return rc, capsys.readouterr().err
+run = lint_runner(main)
 
 
 def skill_doc(description_line: str) -> str:
